+++ passkit_ui/CHANGELOG.md
@@ -1,4 +1,11 @@
+## 0.0.5
+
+- Add ability to create an image from a PkPass
+
 ## 0.0.4
+
+- Improved fidelity in the visual representation of all pass types
+- All transit types have a proper icon
 
 ## 0.0.3